Webapp/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
blinker==1.7.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
docstring_parser==0.16
Flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.1
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.66.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.8
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==2.1.0
packaging==24.1
priority==2.0.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
Quart==0.19.6
quart-cors==0.7.0
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.2.2
vertexai==1.63.0
websockets==13.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
